Create 1.21 LAB 1 (Ex. 1): Check if vector is sorted
Write the inOrder() function, which receives a vector of integers as a parameter, and returns true if the numbers are sorted (in order from low to high) or false otherwise. The program creates 2 vectors. One vector filled with integers provided by the programmer and one vector filled with integers provided by the user. The program outputs whether each vector is in order or not in order. See provided output for first vector to help you with the required output for the second vector.

Ex: If the vector passed to the inOrder() function stores [5, 6, 7, 8, 3], then the function returns false.

Ex: If the vector passed to the inOrder() function stores [4, 6, 7, 8, 9], then the function returns true.