Create 1.22 LAB 1 (Ex. 2): Exact change - functions
Define a function called exactChange that takes the total change amount in cents and an integer vector as parameters. Function exactChange() calculates the change using the fewest coins and stores the number of each coin type used into the vector parameter. Index 0-3 of the vector parameter should contain the number of pennies, nickels, dimes, and quarters respectively. Then write a main program that reads the total change amount as an integer input, calls exactChange(), and outputs the change, one coin type per line. Use singular and plural coin names as appropriate, like 1 penny vs. 2 pennies. Output "no change" if the input is 0 or less.

Ex: If the input is:

0 
(or less), the output is:

no change
Ex: If the input is:

45
the output is:

2 dimes 
1 quarter
Your program must define and call the following function. Positions 0-3 of coinVals should contain the number of pennies, nickels, dimes, and quarters, respectively.
void exactChange(int userTotal, vector<int>& coinVals)