Create O - Sum of integers
Implement the function fileSum. fileSum is passed in a name of a file. This function should open the file, sum all of the integers within this file, close the file, and then return the sum.

If the file does not exist, this function should output an error message and then call the exit function to exit the program passing an error value of EXIT_FAILURE. The exit function and value EXIT_FAILURE is provided by the cstdlib library. Here is how you call the exit function.

exit(EXIT_FAILURE);